TO-407: Support no branch option in jiri patch

+++ cmd/jiri/patch.go
@@ -20,13 +20,14 @@ import (
 )
 
 var (
-	patchRebaseFlag bool
-	patchTopicFlag  bool
-	patchBranchFlag string
-	patchDeleteFlag bool
-	patchHostFlag   string
-	patchForceFlag  bool
-	cherryPickFlag  bool
+	patchRebaseFlag  bool
+	patchTopicFlag   bool
+	patchBranchFlag  string
+	patchDeleteFlag  bool
+	patchHostFlag    string
+	patchForceFlag   bool
+	cherryPickFlag   bool
+	detachedHeadFlag bool
 )
 
 func init() {
@@ -37,6 +38,7 @@ func init() {
 	cmdPatch.Flags.StringVar(&patchHostFlag, "host", "", `Gerrit host to use. Defaults to gerrit host specified in manifest.`)
 	cmdPatch.Flags.BoolVar(&patchTopicFlag, "topic", false, `Patch whole topic.`)
 	cmdPatch.Flags.BoolVar(&cherryPickFlag, "cherry-pick", false, `Cherry-pick patches instead of checking out.`)
+	cmdPatch.Flags.BoolVar(&detachedHeadFlag, "no-branch", false, `Don't create the branch for the patch.`)
 }
 
 // cmdPatch represents the "jiri patch" command.
@@ -57,7 +59,7 @@ flag will delete the branch if already exists. Use the -force flag to force
 deleting the branch even if it contains unmerged changes).
 
 if -topic flag is true jiri will fetch whole topic and will try to apply to
-indivisual projects. Patch will assume topic is of form {USER}-{BRANCH} and
+individual projects. Patch will assume topic is of form {USER}-{BRANCH} and
 will try to create branch name out of it. If this fails default branch name
 will be same as topic. Currently patch does not support the scenario when
 change "B" is created on top of "A" and both have same topic.
@@ -68,41 +70,45 @@ change "B" is created on top of "A" and both have same topic.
 
 // patchProject checks out the given change.
 func patchProject(jirix *jiri.X, local project.Project, ref, branch, remote string) (bool, error) {
-	if branch == "" {
-		cl, ps, err := gerrit.ParseRefString(ref)
-		if err != nil {
-			return false, err
-		}
-		branch = fmt.Sprintf("change/%v/%v", cl, ps)
-	}
-	jirix.Logger.Infof("Patching project %s(%s) on branch %q\n", local.Name, local.Path, branch)
 	scm := gitutil.New(jirix, gitutil.RootDirOpt(local.Path))
 	g := git.NewGit(local.Path)
-	branchExists, err := g.BranchExists(branch)
-	if err != nil {
-		return false, err
-	}
-	if branchExists {
-		if patchDeleteFlag {
-			_, currentBranch, err := g.GetBranches()
+	if !detachedHeadFlag {
+		if branch == "" {
+			cl, ps, err := gerrit.ParseRefString(ref)
 			if err != nil {
 				return false, err
 			}
-			if currentBranch == branch {
-				if err := scm.CheckoutBranch("remotes/origin/"+remote, gitutil.DetachOpt(true)); err != nil {
+			branch = fmt.Sprintf("change/%v/%v", cl, ps)
+		}
+		jirix.Logger.Infof("Patching project %s(%s) on branch %q\n", local.Name, local.Path, branch)
+		branchExists, err := g.BranchExists(branch)
+		if err != nil {
+			return false, err
+		}
+		if branchExists {
+			if patchDeleteFlag {
+				_, currentBranch, err := g.GetBranches()
+				if err != nil {
 					return false, err
 				}
-			}
-			if err := scm.DeleteBranch(branch, gitutil.ForceOpt(patchForceFlag)); err != nil {
-				jirix.Logger.Errorf("Cannot delete branch %q: %s", branch, err)
+				if currentBranch == branch {
+					if err := scm.CheckoutBranch("remotes/origin/"+remote, gitutil.DetachOpt(true)); err != nil {
+						return false, err
+					}
+				}
+				if err := scm.DeleteBranch(branch, gitutil.ForceOpt(patchForceFlag)); err != nil {
+					jirix.Logger.Errorf("Cannot delete branch %q: %s", branch, err)
+					jirix.IncrementFailures()
+					return false, nil
+				}
+			} else {
+				jirix.Logger.Errorf("Branch %q already exists in project %q", branch, local.Name)
 				jirix.IncrementFailures()
 				return false, nil
 			}
-		} else {
-			jirix.Logger.Errorf("Branch %q already exists in project %q", branch, local.Name)
-			jirix.IncrementFailures()
-			return false, nil
 		}
+	} else {
+		jirix.Logger.Infof("Patching project %s(%s)\n", local.Name, local.Path)
 	}
 	if err := scm.FetchRefspec("origin", ref); err != nil {
 		return false, err
@@ -120,13 +126,17 @@ func patchProject(jirix *jiri.X, local project.Project, ref, branch, remote stri
 		}
 		branchBase = "HEAD"
 	}
-	if err := g.CreateBranchFromRef(branch, branchBase); err != nil {
-		return false, err
-	}
-	if err := g.SetUpstream(branch, "origin/"+remote); err != nil {
-		return false, fmt.Errorf("setting upstream to 'origin/%s': %s", remote, err)
-	}
-	if err := scm.CheckoutBranch(branch); err != nil {
+	if !detachedHeadFlag {
+		if err := g.CreateBranchFromRef(branch, branchBase); err != nil {
+			return false, err
+		}
+		if err := g.SetUpstream(branch, "origin/"+remote); err != nil {
+			return false, fmt.Errorf("setting upstream to 'origin/%s': %s", remote, err)
+		}
+		if err := scm.CheckoutBranch(branch); err != nil {
+			return false, err
+		}
+	} else if err := scm.CheckoutBranch(branchBase); err != nil {
 		return false, err
 	}
 	if cherryPickFlag {
